feat(turmas): criação da tela de update e da mudança da bar
