+++ br_ipti_tag_app/lib/app/features/classroom/classroom_module.dart
@@ -7,6 +7,8 @@ import 'package:br_ipti_tag_app/app/features/classroom/domain/usecases/update_cl
 import 'package:br_ipti_tag_app/app/features/classroom/presentation/create/bloc/classroom_create_bloc.dart';
 import 'package:br_ipti_tag_app/app/features/classroom/presentation/update_delete/bloc/classroom_update_bloc.dart';
 import 'package:br_ipti_tag_app/app/features/classroom/presentation/update_delete/update_delete_classroom_page.dart';
+import 'package:br_ipti_tag_app/app/features/classroom/presentation/update_delete/update_student_page.dart';
+import 'package:br_ipti_tag_app/app/features/classroom/presentation/update_delete/update_teacher_page.dart';
 import 'package:flutter_modular/flutter_modular.dart';
 import 'presentation/create/classroom_create_page.dart';
 import 'presentation/list/bloc/classroom_list_bloc.dart';
@@ -46,11 +48,13 @@ class ClassroomModule extends Module {
   final List<ModularRoute<void>> routes = [
     ChildRoute("/", child: (_, args) => const ClassroomPage()),
     ChildRoute("create", child: (_, args) => const ClassroomCreatePage()),
-    ChildRoute("updateDelete",
+    ChildRoute("updatePage",
         child: (_, args) => const ClassroomUpdateDeletePage()),
-    ChildRoute("C",
-        child: (_, args) => const ClassroomUpdateDeletePage()),
-    ChildRoute("UpdateDelete",
-        child: (_, args) => const ClassroomUpdateDeletePage()),
+    ChildRoute("updateTeacher",
+        child: (_, args) => const ClassroomTeacherPage()),
+    ChildRoute("updateStudent",
+        child: (_, args) => const ClassroomStudentPage()),
+    ChildRoute("updateInfos",
+        child: (_, args) => const ClassroomBasicDataForm()),
   ];
 }
